// src/utils/api/types/enum.ts
// LevelIndex 枚举
export enum LevelIndex {
  BASIC = 0,
  ADVANCED = 1,
  EXPERT = 2,
  MASTER = 3,
  ReMASTER = 4,
}

// FCType 枚举
export enum FCType {
  AP_PLUS = "app",
  AP = "ap",
  FC_PLUS = "fcp",
  FC = "fc",
  NONE = "none",
}

// FSType 枚举
export enum FSType {
  FDX_PLUS = "fsdp",
  FDX = "fsd",
  FS_PLUS = "fsp",
  FS = "fs",
  SYNC_PLAY = "sync",
  NONE = "none",
}

// RateType 枚举
export enum RateType {
  SSS_PLUS = "sssp",
  SSS = "sss",
  SS_PLUS = "ssp",
  SS = "ss",
  S_PLUS = "sp",
  S = "s",
  AAA = "aaa",
  AA = "aa",
  A = "a",
  BBB = "bbb",
  BB = "bb",
  B = "b",
  C = "c",
  D = "d",
}

// SongType 枚举
export enum SongType {
  STANDARD = "SD",
  DX = "DX",
  UTAGE = "UT",
}

// 定义 ApiType 枚举
export enum ApiType {
  LXNS = "LXNS",
  FISH = "FISH",
}